+++ glo/trunk/src/Texture2D.cpp
@@ -30,7 +30,7 @@ const bool Texture2D::isBound() const
 	GLint currentBinding;
 	glGetIntegerv( GL_TEXTURE_BINDING_2D, &currentBinding );
 
-	return ( currentBinding == m_object );
+	return ( static_cast<GLuint>(currentBinding) == m_object );
 }
